+++ resize/resize2fs.c
@@ -177,9 +177,9 @@ errcode_t resize_fs(ext2_filsys fs, blk64_t *new_size, int flags,
 #ifdef RESIZE2FS_DEBUG
 	if (rfs->flags & RESIZE_DEBUG_BMOVE)
 		printf("Number of free blocks: %llu/%llu, Needed: %llu\n",
-		       ext2fs_free_blocks_count(rfs->old_fs->super),
-		       ext2fs_free_blocks_count(rfs->new_fs->super),
-		       rfs->needed_blocks);
+		       (unsigned long long) ext2fs_free_blocks_count(rfs->old_fs->super),
+		       (unsigned long long) ext2fs_free_blocks_count(rfs->new_fs->super),
+		       (unsigned long long) rfs->needed_blocks);
 #endif
 
 	init_resource_track(&rtrack, "block_mover", fs->io);
@@ -758,7 +758,7 @@ retry:
 	new_inodes =(unsigned long long) fs->super->s_inodes_per_group * fs->group_desc_count;
 	if (new_inodes > ~0U) {
 		fprintf(stderr, _("inodes (%llu) must be less than %u\n"),
-				   new_inodes, ~0U);
+			(unsigned long long) new_inodes, ~0U);
 		return EXT2_ET_TOO_MANY_INODES;
 	}
 	fs->super->s_inodes_count = fs->super->s_inodes_per_group *
@@ -1644,7 +1644,8 @@ static errcode_t resize2fs_get_alloc_block(ext2_filsys fs,
 
 #ifdef RESIZE2FS_DEBUG
 	if (rfs->flags & 0xF)
-		printf("get_alloc_block allocating %llu\n", blk);
+		printf("get_alloc_block allocating %llu\n",
+		       (unsigned long long) blk);
 #endif
 
 	ext2fs_mark_block_bitmap2(rfs->old_fs->block_map, blk);
@@ -1750,7 +1751,9 @@ static errcode_t block_mover(ext2_resize_t rfs)
 #ifdef RESIZE2FS_DEBUG
 		if (rfs->flags & RESIZE_DEBUG_BMOVE)
 			printf("Moving %llu blocks %llu->%llu\n",
-			       size, old_blk, new_blk);
+			       (unsigned long long) size,
+			       (unsigned long long) old_blk,
+			       (unsigned long long) new_blk);
 #endif
 		do {
 			c = size;
@@ -1843,8 +1846,9 @@ static int process_block(ext2_filsys fs, blk64_t	*block_nr,
 #ifdef RESIZE2FS_DEBUG
 			if (pb->rfs->flags & RESIZE_DEBUG_BMOVE)
 				printf("ino=%u, blockcnt=%lld, %llu->%llu\n",
-				       pb->old_ino, blockcnt, block,
-				       new_block);
+				       pb->old_ino, (long long) blockcnt,
+				       (unsigned long long) block,
+				       (unsigned long long) new_block);
 #endif
 			block = new_block;
 		}
@@ -2460,7 +2464,8 @@ static errcode_t move_itables(ext2_resize_t rfs)
 #ifdef RESIZE2FS_DEBUG
 		if (rfs->flags & RESIZE_DEBUG_ITABLEMOVE)
 			printf("Itable move group %d block %llu->%llu (diff %lld)\n",
-			       i, old_blk, new_blk, diff);
+			       i, (unsigned long long) old_blk,
+			       (unsigned long long) new_blk, diff);
 #endif
 
 		if (!diff)
@@ -2934,7 +2939,8 @@ blk64_t calculate_minimum_resize_size(ext2_filsys fs, int flags)
 	}
 #ifdef RESIZE2FS_DEBUG
 	if (flags & RESIZE_DEBUG_MIN_CALC)
-		printf("fs requires %llu data blocks.\n", data_needed);
+		printf("fs requires %llu data blocks.\n",
+		       (unsigned long long) data_needed);
 #endif
 
 	/*
@@ -2977,7 +2983,7 @@ blk64_t calculate_minimum_resize_size(ext2_filsys fs, int flags)
 #ifdef RESIZE2FS_DEBUG
 	if (flags & RESIZE_DEBUG_MIN_CALC)
 		printf("With %d group(s), we have %llu blocks available.\n",
-		       groups, data_blocks);
+		       groups, (unsigned long long) data_blocks);
 #endif
 
 	/*
@@ -3030,8 +3036,10 @@ blk64_t calculate_minimum_resize_size(ext2_filsys fs, int flags)
 		if (flags & RESIZE_DEBUG_MIN_CALC)
 			printf("Added %d extra group(s), "
 			       "blks_needed %llu, data_blocks %llu, "
-			       "last_start %llu\n", extra_grps, blks_needed,
-			       data_blocks, last_start);
+			       "last_start %llu\n", extra_grps,
+			       (unsigned long long) blks_needed,
+			       (unsigned long long) data_blocks,
+			       (unsigned long long) last_start);
 #endif
 	}
 
@@ -3046,7 +3054,8 @@ blk64_t calculate_minimum_resize_size(ext2_filsys fs, int flags)
 
 #ifdef RESIZE2FS_DEBUG
 	if (flags & RESIZE_DEBUG_MIN_CALC)
-		printf("Last group's overhead is %llu\n", overhead);
+		printf("Last group's overhead is %llu\n",
+		       (unsigned long long) overhead);
 #endif
 
 	/*
@@ -3059,7 +3068,7 @@ blk64_t calculate_minimum_resize_size(ext2_filsys fs, int flags)
 #ifdef RESIZE2FS_DEBUG
 		if (flags & RESIZE_DEBUG_MIN_CALC)
 			printf("Need %llu data blocks in last group\n",
-			       remainder);
+			       (unsigned long long) remainder);
 #endif
 		/*
 		 * 50 is a magic number that mkfs/resize uses to see if its
@@ -3077,7 +3086,8 @@ blk64_t calculate_minimum_resize_size(ext2_filsys fs, int flags)
 	overhead += fs->super->s_first_data_block;
 #ifdef RESIZE2FS_DEBUG
 	if (flags & RESIZE_DEBUG_MIN_CALC)
-		printf("Final size of last group is %lld\n", overhead);
+		printf("Final size of last group is %llu\n",
+		       (unsigned long long) overhead);
 #endif
 
 	/* Add extra slack for bigalloc file systems */
@@ -3104,7 +3114,8 @@ blk64_t calculate_minimum_resize_size(ext2_filsys fs, int flags)
 
 #ifdef RESIZE2FS_DEBUG
 	if (flags & RESIZE_DEBUG_MIN_CALC)
-		printf("Estimated blocks needed: %llu\n", blks_needed);
+		printf("Estimated blocks needed: %llu\n",
+		       (unsigned long long) blks_needed);
 #endif
 
 	/*
@@ -3139,7 +3150,8 @@ blk64_t calculate_minimum_resize_size(ext2_filsys fs, int flags)
 
 #ifdef RESIZE2FS_DEBUG
 		if (flags & RESIZE_DEBUG_MIN_CALC)
-			printf("Extents safety margin: %llu\n", safe_margin);
+			printf("Extents safety margin: %llu\n",
+			       (unsigned long long) safe_margin);
 #endif
 		blks_needed += safe_margin;
 	}
